+++ src/werkzeug/sansio/request.py
@@ -27,6 +27,8 @@ from ..http import parse_list_header
 from ..http import parse_options_header
 from ..http import parse_range_header
 from ..http import parse_set_header
+from ..http import SecFetchMode
+from ..http import SecFetchSite
 from ..user_agent import UserAgent
 from ..utils import cached_property
 from ..utils import header_property
@@ -519,6 +521,25 @@ class Request:
             " on the response to indicate which methods are allowed."
         ),
         read_only=True,
+    )
+
+    sec_fetch_mode = header_property[SecFetchMode](
+        "Sec-Fetch-Mode",
+        load_func=lambda value: SecFetchMode(value),
+        dump_func=lambda value: value.value,
+        doc="""Allows identification of requests originating from a
+        user navigating between HTML pages, and requests to load
+        images and other resources. Values must be a member of the
+        :class:`werkzeug.http.SecFetchMode` enum.""",
+    )
+
+    sec_fetch_site = header_property[SecFetchSite](
+        "Sec-Fetch-Site",
+        load_func=lambda value: SecFetchSite(value),
+        dump_func=lambda value: value.value,
+        doc="""Indicates the relationship between a request initiator's
+        origin and the origin of the requested resource. Values must be
+        a member of the :class:`werkzeug.http.SecFetchSite` enum.""",
     )
 
     @property